backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.14
django-cors-headers>=4.3
psycopg2-binary>=2.9
dj-database-url>=2.1
python-dotenv>=1.0
requests>=2.28
